Fix pg_autoctl config set pg_autoctl.name or hostname.
We need to take the new value in the local configuration file (that was ok)
and also update the monitor with the new value (that was not done).

Note that because of a change in the API where we need the whole keeper, the
function keeper_config_accept_new must move to the keeper.c file ("module").